ADD current commute data API

+++ src/main/java/com/project/noris/PCutilization/dto/TeamdataDto.java
@@ -18,8 +18,8 @@ public class TeamdataDto {
 
     private double percent;
 
-    private int total_time;
+    private double total_time;
 
-    private int work_time;
+    private double work_time;
 
 }
